Introduces formal_value function from shap to have identical looking plots. Small Refactor of force
